Match

Documentation

data Match

Tipo de dato Match que se encarga de Clasificar las letras en Toro, Vaca, Nada

Constructors

Nada

Vaca

Toro

checkGuess

Arguments

:: [Char]
Palabra que es la adivinación del usuario.

-> [Char]
Palabra elegida por el juego que el usuario debe adivinar

-> [Match]
Arrglo de vaca,toros o guiones para saber si se adivino o no la palabra

Funciones para llamar todo el proceso de verificación de la adivinación. Crea dos tipos de arreglos de pares (Char,Match) y (Char,Bool) respectivamente.

lookToros

Arguments

:: [(Char, Match)]
Dupla de las letras de la adivinación con Nada

-> [(Char, Bool)]
Dupla del objetivo a adivinar con False

-> [(Char, Match)]
Dupla de las letras de la adivinación, con Toro si es que coincidio con el objetivo

-> [(Char, Bool)]
Dupla del objetivo a adivinar con True, si es que coincidio con la adivinación

-> [Match]
Arreglo de Match que dicen que tan cerca esta la adivinación del Objetivo

Función que se encarga de verificar los toros de la adivinación. Recibe 4 argumentos arreglos de pares (Char,Match), (Char,Bool) y dos arreglos vacios que se llenaran mientras hace la recursión. Si una letra de la dupla de Match en la posición i-esima coincide con su homologo de la dupla de Bool, se guarda en el primer arreglo vació la convinación (Char,Toro) y en el segundo arreglo vacio (Char,True), para saber que esa letra esta en buena posición y que la primera letra del objetivo ya fue usada.

lookVacas :: Eq a => [(a, Match)] -> [(a, Bool)] -> [Match]

Esta función se encarga de encontrar las vacas entre la adivinación y el objetivo. Recibe dos arreglos de duplas, uno de (Char,Match) y otro de (Char,Bool) respectivamente. Procede a comparar cada char del primer arreglo con cada char del segundo.

lectura :: [Char] -> [Match]

Función auxiliar para poder leer un arreglo de match