Play

Documentation

playGame :: IO ()

Función que inicia el juego, carga todas las palabras, y elige una al azar.

gameStatus

Arguments

:: [Char]
Palabra seleccionada a Azar

-> Int
Numero de turnos que lleva el juego

-> IO ()
No retorna nada. Ya que todo se maneja con prints

El ciclo del juego, llama a las funciones de verificación de la adivinación y se encarga de terminar el juego, de acertar la palabra o despues de los turnos estipulados.

leerEntrada :: IO String

Función Dedicada a leer la entrada del usuario, manteniendo el ciclo hasta que de una entrada Correcta.