Solve

Documentation

playSolver :: IO ()

Función a Exportar y que Inicia el estado del SOLVER Carga todas las palabras en una lista, setea el turno en 0 y elige una palabra al azar para iniciar

solveTheGame :: String -> Int -> [String] -> [String] -> IO ()

Función que maneja el flujo del solver, muestra la palabra sugerida, y luego pide la pista. Si la pista no es completamente buena, procede a buscar una palabra a sugerir y se repite el proceso. Si se llega a más 6 turnos. Finaliza el programa.

leerEntradaSolver :: IO [Match]

Función encargada de leer la entrada, verificar si es posible dicha entrada. Y luego convertirla en un tipo de dato Match, que es el codigo utilizado para las palabras.